In progress of making the shop module fully functional, including all of the (tedious) CRUD stuff in both the module API as the REST API

+++ shopping-api/src/main/java/com/programyourhome/shop/model/PyhProductAggregation.java
@@ -1,5 +1,6 @@
 package com.programyourhome.shop.model;
 
+import java.math.BigDecimal;
 import java.util.Collection;
 
 public interface PyhProductAggregation {
@@ -15,14 +16,14 @@ public interface PyhProductAggregation {
      *
      * @return the minimum amount
      */
-    public int getMinimumAmount();
+    public BigDecimal getMinimumAmount();
 
     /**
      * Get the maximum amount of items that should be 'available' of this product aggregation.
      *
      * @return the maximum amount
      */
-    public int getMaximumAmount();
+    public BigDecimal getMaximumAmount();
 
     public Collection<? extends PyhProductAggregationPart> getAggregationParts();
